+++ exercises/hamming/src/example.h
@@ -1,6 +1,6 @@
 #ifndef _HAMMING_H
 #define _HAMMING_H
 
-int compute(const char *lhs, const char *rhs);
+size_t compute(const char *lhs, const char *rhs);
 
 #endif
